requirements.txt
anyio==3.6.1
APScheduler==3.9.1
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
loguru==0.6.0
lxml==4.9.1
multiprocess==0.70.13
mypy-extensions==0.4.3
pathos==0.2.9
pathspec==0.9.0
platformdirs==2.5.2
pox==0.3.1
ppft==1.7.6.5
PyExecJS==1.5.1
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tomli==2.0.1
tornado==6.2
translators==5.4.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
win32-setctime==1.1.0
